geninv_codec/requirements.txt
numpy
pillow
torch
matplotlib
